refactored folder names , renamed for better understanding and file structure
