+++ src/controllers/user-actions/workspace-actions/messageController.js
@@ -1,5 +1,5 @@
-const Message = require("../../models/Message");
-const asyncErrorHandler = require("../../utils/AsyncErrorHandler");
+const Message = require("../../../models/Message");
+const asyncErrorHandler = require("../../../utils/AsyncErrorHandler");
 
 // MESSAGE SECTION
 exports.getChannelMessageById = asyncErrorHandler(async (req, res, next) => {
