+++ src/router/user-actions/workspace-routes/inviteRoute.js
@@ -1,6 +1,6 @@
 const express = require("express");
-const inviteController = require("../../controllers/user-actions/inviteController");
-const authController = require("../../controllers/user-auth/authController");
+const inviteController = require("../../../controllers/user-actions/workspace-actions/inviteController");
+const authController = require("../../../controllers/user-auth/authController");
 
 const router = express.Router();
